requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dateutils==0.6.12
dnspython==2.6.1
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openpyxl==3.1.2
orjson==3.9.14
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
tweety-ns==1.0.9.5
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
